+++ apps/uiservice/src/app/(auth)/products/httpnosql/page.tsx
@@ -17,8 +17,8 @@ export default function Page() {
     return (
       <TableRow className="cursor-pointer" key={data._id}>
         <TableCell><div className="font-medium">{data?.key}</div></TableCell>
-        <TableCell className="text-slate-500"><DataModal dataObj={data} /></TableCell>
-        <TableCell className="text-right text-slate-500 hidden md:table-cell">{format(new Date(data.createdAt), "MMM, do yyyy, h:mm a")}</TableCell>
+        <TableCell className="text-zinc-500"><DataModal dataObj={data} /></TableCell>
+        <TableCell className="text-right text-zinc-500 hidden md:table-cell">{format(new Date(data.createdAt), "MMM, do yyyy, h:mm a")}</TableCell>
       </TableRow>
     )
   })