+++ apps/uiservice/src/app/page.tsx
@@ -23,13 +23,13 @@ export default function Page() {
             <h2 className="font-bold text-lg capitalize">{tier.computeTier} Tier</h2>
             <ul className="grid gap-3 text-sm text-muted-foreground">
               {Object.entries(tier.estimatedRequestCost).map(([key, value]) => (
-                <li className="flex text-xs items-center text-slate-600" key={key}>
+                <li className="flex text-xs items-center text-zinc-600" key={key}>
                   <CheckCircle2 className="scale-75 me-2" />
                   {brandName} {products?.data?.find((item: any) => item?.productName === key)?.displayName}
                   {" "}$ {Number(value).toFixed(2)}/req
                 </li>
               ))}
-              <li className="flex text-xs items-center text-slate-600">
+              <li className="flex text-xs items-center text-zinc-600">
                 <CheckCircle2 className="scale-75 me-2" /> {tier.responseDelay} ms response delay
               </li>
             </ul>
@@ -46,7 +46,7 @@ export default function Page() {
           <div dangerouslySetInnerHTML={{ __html: product?.productIcon }} style={{ zoom: "150%" }}></div>
           <div className="space-y-2">
             <h3 className="font-bold">{brandName} {product?.displayName}</h3>
-            <p className="text-sm text-slate-600">
+            <p className="text-sm text-zinc-600">
               {product?.description}
             </p>
           </div>
@@ -62,7 +62,7 @@ export default function Page() {
           <div dangerouslySetInnerHTML={{ __html: solution?.solutionIcon }} style={{ zoom: "150%" }}></div>
           <div className="space-y-2">
             <h3 className="font-bold">{solution?.solutionName}</h3>
-            <p className="text-sm text-slate-600">
+            <p className="text-sm text-zinc-600">
               {solution?.description}
             </p>
           </div>
@@ -96,12 +96,12 @@ export default function Page() {
             </div>
           </div>
         </section>
-        <section id="solutions" className="mt-8 container space-y-6 bg-slate-50 py-8 dark:bg-transparent md:py-12 lg:py-24 lg:rounded-lg">
+        <section id="solutions" className="mt-8 container space-y-6 bg-zinc-50 py-8 dark:bg-transparent md:py-12 lg:py-24 lg:rounded-lg">
           <div className="mx-auto flex max-w-[58rem] flex-col items-center space-y-4 text-center">
             <h2 className="font-heading text-3xl leading-[1.1] sm:text-3xl md:text-5xl">
               Solutions
             </h2>
-            <p className="max-w-[85%] leading-normal text-slate-600 sm:text-lg sm:leading-7">
+            <p className="max-w-[85%] leading-normal text-zinc-600 sm:text-lg sm:leading-7">
               {uiConstants.solutionHeader}
             </p>
           </div>
@@ -114,7 +114,7 @@ export default function Page() {
             <h2 className="font-heading text-3xl leading-[1.1] sm:text-3xl md:text-6xl">
               Open Source
             </h2>
-            <p className="max-w-[85%] leading-normal text-slate-600 sm:text-lg sm:leading-7">
+            <p className="max-w-[85%] leading-normal text-zinc-600 sm:text-lg sm:leading-7">
               {uiConstants.openSourceHeader} <br />{" "}
               The code is available on{" "}
               <Link
@@ -145,12 +145,12 @@ export default function Page() {
             </Link>
           </div>
         </section>
-        <section id="products" className="container space-y-6 bg-slate-50 py-8 dark:bg-transparent md:py-12 lg:py-24 lg:rounded-lg">
+        <section id="products" className="container space-y-6 bg-zinc-50 py-8 dark:bg-transparent md:py-12 lg:py-24 lg:rounded-lg">
           <div className="mx-auto flex max-w-[58rem] flex-col items-center space-y-4 text-center">
             <h2 className="font-heading text-3xl leading-[1.1] sm:text-3xl md:text-5xl">
               Products
             </h2>
-            <p className="max-w-[85%] leading-normal text-slate-600 sm:text-lg sm:leading-7">
+            <p className="max-w-[85%] leading-normal text-zinc-600 sm:text-lg sm:leading-7">
               {uiConstants.productsHeader}
             </p>
           </div>
@@ -163,7 +163,7 @@ export default function Page() {
             <h2 className="font-heading text-3xl leading-[1.1] sm:text-3xl md:text-5xl">
               Compute Tiers
             </h2>
-            <p className="max-w-[85%] leading-normal text-slate-600 sm:text-lg sm:leading-7">
+            <p className="max-w-[85%] leading-normal text-zinc-600 sm:text-lg sm:leading-7">
               {uiConstants.computeTierHeader}
             </p>
           </div>
@@ -173,7 +173,7 @@ export default function Page() {
         </section>
       </div >
       <footer>
-        <div className="bg-slate-50">
+        <div className="bg-zinc-50">
           <div className="container flex flex-col items-center justify-between gap-4 py-10 md:h-24 md:flex-row md:py-0">
             <div className="flex flex-col items-center gap-4 px-8 md:flex-row md:gap-2 md:px-0">
               <p className="text-center text-sm leading-loose md:text-left">